Mass renames.
