+++ src/app/codegen-data-model/CodegenDataModel.cpp
@@ -14,7 +14,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-#include <app/codegen-interaction-model/CodegenDataModel.h>
+#include <app/codegen-data-model/CodegenDataModel.h>
 
 #include <app-common/zap-generated/attribute-type.h>
 #include <app/RequiredPrivilege.h>
