+++ src/app/codegen-data-model/CodegenDataModel.h
@@ -16,7 +16,7 @@
  */
 #pragma once
 
-#include <app/interaction-model/Model.h>
+#include <app/data-model-interface/DataModel.h>
 
 #include <app/util/af-types.h>
 
@@ -35,7 +35,7 @@ namespace app {
 /// Given that this relies on global data at link time, there generally can be
 /// only one CodegenDataModel per application (you can create more instances,
 /// however they would share the exact same underlying data and storage).
-class CodegenDataModel : public chip::app::InteractionModel::Model
+class CodegenDataModel : public chip::app::InteractionModel::DataModel
 {
 private:
     /// Ember commands are stored as a `CommandId *` pointer that is either null (i.e. no commands)
