+++ src/app/codegen-data-model/model.gni
@@ -25,11 +25,11 @@ import("//build_overrides/chip.gni")
 # be cleanly built as a stand-alone and instead have to be imported as part of
 # a different data model or compilation unit.
 codegen_interaction_model_SOURCES = [
-  "${chip_root}/src/app/codegen-interaction-model/CodegenDataModel.h",
-  "${chip_root}/src/app/codegen-interaction-model/CodegenDataModel.cpp",
+  "${chip_root}/src/app/codegen-data-model/CodegenDataModel.h",
+  "${chip_root}/src/app/codegen-data-model/CodegenDataModel.cpp",
 ]
 
 codegen_interaction_model_PUBLIC_DEPS = [
   "${chip_root}/src/app/common:attribute-type",
-  "${chip_root}/src/app/interaction-model",
+  "${chip_root}/src/app/data-model-interface",
 ]
