+++ src/app/data-model-interface/tests/BUILD.gn
@@ -13,23 +13,13 @@
 # limitations under the License.
 import("//build_overrides/chip.gni")
 import("${chip_root}/build/chip/chip_test_suite.gni")
-import("${chip_root}/src/app/codegen-interaction-model/model.gni")
-
-source_set("mock_model") {
-  sources = codegen_interaction_model_SOURCES
-
-  public_deps = codegen_interaction_model_PUBLIC_DEPS
-
-  # this ties in the codegen model to an actual ember implementation
-  public_deps += [ "${chip_root}/src/app/util/mock:mock_ember" ]
-}
 
 chip_test_suite("tests") {
-  output_name = "libCodegenInteractionModelTests"
+  output_name = "libIMInterfaceTests"
 
-  test_sources = [ "TestCodegenModelViaMocks.cpp" ]
+  test_sources = [ "TestEventEmitting.cpp" ]
 
   cflags = [ "-Wconversion" ]
 
-  public_deps = [ ":mock_model" ]
+  public_deps = [ "${chip_root}/src/app/data-model-interface" ]
 }
